[Refactor] Extract network-receive-system
